requirements.txt
streamlit>=1.40.0

# LangChain Core
langchain
langchain-core
langchain_community

# OpenAI Integration (optional, used in some demos)
langchain-openai
openai

# Anthropic Integration (optional)
langchain-anthropic

# Google Gemini Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration (used in ai_suggester.py)
langchain-huggingface
transformers
huggingface-hub
sentence-transformers

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# PyTorch (only needed for some Hugging Face models)
torch
torchvision
torchaudio